// src/components/models/invoice-types.tsx
export type Invoice = {
  id: string;
  invoice_status: string;
  invoice_date: string;
  payment_terms: string;
  project_description: string;
  street_address: string;
  city: string;
  post_code: string;
  country: string;
  clients_name: string;
  clients_email: string;
  clients_street_address: string;
  clients_city: string;
  clients_post_code: string;
  clients_country: string;
  item_name: string;
  quantity: number;
  price: number;
};